remove references to removed code
